refactor(registers): simplified class structure regarding register
